refactor(mcp): extract streamable HTTP transport
